+++ Sources/Parser.swift
@@ -20,62 +20,6 @@
 import Foundation
 
 /**
- Used to represent whether the parsing operation was successful or encountered 
- an error.
- */
-public enum Result<SuccessType> {
-    /**
-     Indicates the parsing operation was successful and resulted in the
-     provided associated value.
-     */
-    case success(SuccessType)
-
-    /**
-     Indicates the parsing operation failed and resulted in the provided
-     associated error value.
-     */
-    case failure(Error)
-
-    /// Returns `true` if the result is a success, `false` otherwise.
-    public var isSuccess: Bool {
-        switch self {
-        case .success:
-            return true
-        case .failure:
-            return false
-        }
-    }
-
-    /// Returns `true` if the result is a failure, `false` otherwise.
-    public var isFailure: Bool {
-        return !isSuccess
-    }
-
-    /**
-     Returns the assoicated value if the result is a success, `nil` otherwise.
-     */
-    public var value: SuccessType? {
-        if case let .success(value) = self {
-            return value
-        }
-
-        return nil
-    }
-
-    /**
-     Returns the associated error value if the result is a failure, `nil` 
-     otherwise.
-     */
-    public var error: Error? {
-        if case let .failure(error) = self {
-            return error
-        }
-
-        return nil
-    }
-}
-
-/**
  Instances of this class parse XML documents into a tree of DOM objects.
  */
 public class Parser {
@@ -164,12 +108,14 @@ public class Parser {
         self.parser = parser
     }
 
+    public typealias ParserResult = Result<Document>
+
     /**
      Performs the parsing operation.
      
      - returns: The result of the parsing operation.
      */
-    public func parse() -> Result<Document> {
+    public func parse() -> ParserResult {
         let stack = NodeStack()
         parser.delegate = stack
